requirements.txt
aiogram~=3.0
environs~=9.0
redis
betterlogging~=1.0.0

tabula-py

# # For enabling api:
# backoff
# ujson
# yarl

# # For PostgreSQL sqlalchemy + alembic:
sqlalchemy~=2.0
alembic~=1.0
asyncpg
python-dateutil~=2.9.0.post0
pandas~=2.2.2
matplotlib
